prakhar/Storage_Dashboard/Storage_Dashboard/requirements.txt
Babel==2.11.0
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-Babel==2.0.0
Flask-MySQL==1.5.2
Flask-Table==0.5.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.31
numpy==1.19.5
pandas==1.1.5
protobuf==3.19.6
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
